application updated

+++ src/main/java/com/miu/realestate/repository/ApplicationRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApplicationRepo extends CrudRepository<Application,Integer> {
+public interface ApplicationRepo extends CrudRepository<Application,Long> {
 }
